gh-2055 Extracted UI into base class
