+++ Multisig/Features/Connect to Wallet/UI/SelectWalletViewController.swift
@@ -144,7 +144,7 @@ extension SelectWalletViewController: UITableViewDelegate, UITableViewDataSource
     func connect(to wallet: WCAppRegistryEntry) {
         cancelExistingConnection()
         let chain = Selection.current().safe?.chain ?? Chain.mainnetChain()
-        let walletConnectionVC = WalletConnectionViewController(wallet: wallet, chain: chain)
+        let walletConnectionVC = StartWalletConnectionViewController(wallet: wallet, chain: chain)
         walletConnectionVC.onSuccess = { [weak walletConnectionVC, weak self] connection in
             self?.connection = connection
             walletConnectionVC?.dismiss(animated: true) {
