+++ Multisig/UI/Settings/OwnerKeyManagement/OwnerKeyDetailsViewController/OwnerKeyDetailsViewController.swift
@@ -388,7 +388,7 @@ class OwnerKeyDetailsViewController: UITableViewController {
 
         let chain = Selection.current().safe?.chain ?? Chain.mainnetChain()
 
-        let walletConnectionVC = WalletConnectionViewController(wallet: wcWallet, chain: chain)
+        let walletConnectionVC = StartWalletConnectionViewController(wallet: wcWallet, chain: chain)
         walletConnectionVC.onSuccess = { [weak walletConnectionVC, weak self] connection in
             walletConnectionVC?.dismiss(animated: true) {
                 guard connection.accounts.contains(keyInfo.address) else {
